2D sketch and drawing models
I made these by scaling our plane from the geometry of the Super Tucano tail in accordance to Oliver's tail area calculations from the excel sheet "Tail Sizing Props" and wings of the Stuka. The 2D design therefore is sensitive to that excel sheet because the dimensions I used were sourced from that.

The most useful files in here are the FINAL_TwoViewDWG which we used for our presentation. Other files are components that make up that file like the side view isolated instead of combined etc.
The folder behind the scenes work is scratch work and doesnt need to be accessed unless you are curios about the process I used.